refactor: extract json loading utilities into dedicated module
- Move load_json_file function from file_operations.py to new json_utils.py
- Update imports across cli/handlers.py, core/conversion_strategies.py, and core/converter.py
- Improve code organization by separating JSON-specific utilities from general file operations
- Maintain comprehensive error handling and validation in the new module
